requirements.txt
astroid==2.4.2
boto3==1.14.47
botocore==1.17.47
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
docutils==0.15.2
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.1
python-telegram-bot==12.8
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
